fix: redirect when no user squads
