+++ packages/webapp/components/footer/common.ts
@@ -1,5 +1,5 @@
 import { LoggedUser } from '@dailydotdev/shared/src/lib/user';
-import { ReactElement } from 'react';
+import { ReactElement, ReactNode } from 'react';
 
 export interface FooterTab {
   path?: string | ((user: LoggedUser) => string);
@@ -12,6 +12,7 @@ export interface FooterTab {
 
 export interface FooterNavBarContainerProps {
   activeTab: string;
+  tabs: (FooterTab | ReactNode)[];
 }
 
 export const getNavPath = (
